Create LogSearch.cpp
Here is the source code for my additional functionality of searching the log with a sample main(). The only thing it needs is something added to the beginning of the logSearch() function to check the validity of the item number the user wishes to search, unless we want to check the number before sending it to the logSearch() function.

The additional function and class are necessary for the program to work. I just found it easier to make an additional function instead of putting it in the function. The searchLog() is the only function that should access the other functions and class. I am also using the class in the log creation function. Which I will be uploading shortly.