Create LogAdd.cpp
This is the function to open and add to the log file every time an item is processed.

NOTE: the INorOUT string, this is to display weather the item is being added or removed from the warehouses. This can be changed if someone thinks of a better way to do this... maybe have a bool instead? idk. I'll bring it up at Thursday's meeting.