Update Main.cpp (Log - INCOMPLETE - see description)
Updated logSearch() so that when an item isn't found in the log it will prompt the user to search for another item or exit the search.
Moved the history log related prototypes to before the logSearch() function.
Updated the function call for the history log search main menu option.
Updated the searchHistoryLog() function to prompt the user for an item number to search and check the input.
Fixed an error with the AddToLogFile() function with the quantity. (changed the quantity parameter from integer type to string type)

NOTE: I still need to add the "AddToLogFile()" function calls in the order/shipments processing functions. I will need to review Anna's algorithm or speak with her to determine the proper location for the function calls. I tried inserting them in the load shipments/orders vectors but it did not work correctly. I could not fine the correct location myself and will just need to review the code and algorithms more to figure out what the functions are doing exactly.