Update Main.cpp
Fixed my function. When the user changes the size of a given item, the warehouse now gets updated. The function will copy that item's content, search for an available spot inside the warehouse for the new size, then apply it to that location. After it is copied; deletes the old.